w3c-reports/12092014/mac/runned-tests/XMLHttpRequest/send-non-same-origin.sub-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load mailto:test@example.org. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: XMLHttpRequest cannot load tel:+31600000000. Cross origin requests are only supported for HTTP.
Blocked access to external URL http://www2.localhost:8800/
CONSOLE MESSAGE: XMLHttpRequest cannot load javascript:alert('FAIL'). Cross origin requests are only supported for HTTP.
Blocked access to external URL http://www2.localhost:8800/folder.txt

FAIL XMLHttpRequest: send() - non same-origin (mailto:test@example.org) assert_throws: function "function () { client.send(); }" threw object "Error: NETWORK_ERR: XMLHttpRequest Exception 101" that is not a DOMException NetworkError: property "code" is equal to 101, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:18:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
url@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:15:13
global code@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:21:10)
FAIL XMLHttpRequest: send() - non same-origin (tel:+31600000000) assert_throws: function "function () { client.send(); }" threw object "Error: NETWORK_ERR: XMLHttpRequest Exception 101" that is not a DOMException NetworkError: property "code" is equal to 101, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:18:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
url@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:15:13
global code@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:22:10)
FAIL XMLHttpRequest: send() - non same-origin (http://www2.localhost:8800/) assert_throws: function "function () { client.send(); }" threw object "Error: ABORT_ERR: XMLHttpRequest Exception 102" that is not a DOMException NetworkError: property "code" is equal to 102, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:18:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
url@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:15:13
global code@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:23:10)
FAIL XMLHttpRequest: send() - non same-origin (javascript:alert('FAIL')) assert_throws: function "function () { client.send(); }" threw object "Error: NETWORK_ERR: XMLHttpRequest Exception 101" that is not a DOMException NetworkError: property "code" is equal to 101, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:18:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
url@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:15:13
global code@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:24:10)
FAIL XMLHttpRequest: send() - non same-origin (folder.txt) assert_throws: function "function () { client.send(); }" threw object "Error: ABORT_ERR: XMLHttpRequest Exception 102" that is not a DOMException NetworkError: property "code" is equal to 102, expected 19(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:682:23
http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:18:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
url@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:15:13
global code@http://localhost:8800/XMLHttpRequest/send-non-same-origin.sub.htm:25:10)
